+++ src/main/java/wooteco/subway/line/domain/Sections.java
@@ -5,9 +5,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import wooteco.subway.station.exception.InvalidDeletionException;
-import wooteco.subway.station.exception.NotFoundException;
+import wooteco.subway.exception.NotFoundException;
 import wooteco.subway.station.domain.Station;
+import wooteco.subway.station.exception.InvalidDeletionException;
 
 public class Sections {
     private List<Section> sections = new ArrayList<>();
