+++ src/main/java/wooteco/subway/line/dto/LineRequest.java
@@ -1,6 +1,5 @@
 package wooteco.subway.line.dto;
 
-import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
