+++ src/main/java/wooteco/subway/line/exception/DuplicatedLineException.java
@@ -1,10 +1,7 @@
 package wooteco.subway.line.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import wooteco.subway.line.dto.LineRequest;
 
-@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class DuplicatedLineException extends RuntimeException {
 
     public DuplicatedLineException(LineRequest request) {
