+++ src/main/java/wooteco/subway/member/exception/DuplicatedMemberException.java
@@ -1,9 +1,5 @@
 package wooteco.subway.member.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class DuplicatedMemberException extends RuntimeException {
 
     public DuplicatedMemberException(String message) {
