+++ src/main/java/wooteco/subway/path/exception/InvalidPathException.java
@@ -1,9 +1,5 @@
 package wooteco.subway.path.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class InvalidPathException extends RuntimeException {
 
     public InvalidPathException(String message) {
