+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.exception.NotFoundException;
 import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
@@ -11,7 +12,6 @@ import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 import wooteco.subway.station.exception.DuplicatedStationException;
 import wooteco.subway.station.exception.InvalidDeletionException;
-import wooteco.subway.station.exception.NotFoundException;
 
 @Service
 @Transactional(readOnly = true)
