+++ src/main/java/wooteco/subway/station/exception/DuplicatedStationException.java
@@ -1,9 +1,5 @@
 package wooteco.subway.station.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class DuplicatedStationException extends RuntimeException {
 
     public DuplicatedStationException(String name) {
