+++ src/main/java/wooteco/subway/station/exception/InvalidDeletionException.java
@@ -1,9 +1,5 @@
 package wooteco.subway.station.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class InvalidDeletionException extends RuntimeException {
 
     public InvalidDeletionException(String message) {
