--- src/main/java/wooteco/subway/station/exception/NotFoundException.java
@@ -1,13 +1,0 @@
-package wooteco.subway.station.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-
-}
