+++ src/test/java/wooteco/subway/station/application/StationServiceTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoSettings;
+import wooteco.subway.exception.NotFoundException;
 import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
@@ -20,7 +21,6 @@ import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 import wooteco.subway.station.exception.DuplicatedStationException;
 import wooteco.subway.station.exception.InvalidDeletionException;
-import wooteco.subway.station.exception.NotFoundException;
 
 @MockitoSettings
 class StationServiceTest {
